+++ src/main/java/com/nbrichau/vanillaextension/fences/LogFence.java
@@ -25,7 +25,7 @@ public class LogFence extends FenceBlock {
 			if (!worldIn.isRemote) {
 				worldIn.setBlockState(pos, block.getDefaultState().with(NORTH, state.get(NORTH)).with(EAST, state.get(EAST)).with(SOUTH, state.get(SOUTH)).with(WEST, state.get(WEST)).with(WATERLOGGED, state.get(WATERLOGGED)), 11);
 				if (player != null) {
-					player.getHeldItem(handIn).damageItem(1, player, item->item.sendBreakAnimation(handIn));
+					player.getHeldItem(handIn).damageItem(1, player, item -> item.sendBreakAnimation(handIn));
 				}
 				return ActionResultType.SUCCESS;
 			}
@@ -33,10 +33,7 @@ public class LogFence extends FenceBlock {
 		return ActionResultType.PASS;
 	}
 
-	private static Block getBlockStripping(Block blockIn){
-		System.out.println(blockIn.getRegistryName().getNamespace());
-		System.out.println(blockIn.getRegistryName().getPath());
-		System.out.println(ResourceLocation.create(blockIn.getRegistryName().getNamespace()+":stripped_"+blockIn.getRegistryName().getPath(),':'));
-		return ForgeRegistries.BLOCKS.getValue(ResourceLocation.create(blockIn.getRegistryName().getNamespace()+":stripped_"+blockIn.getRegistryName().getPath(),':'));
+	private static Block getBlockStripping(Block blockIn) {
+		return ForgeRegistries.BLOCKS.getValue(ResourceLocation.create(blockIn.getRegistryName().getNamespace() + ":stripped_" + blockIn.getRegistryName().getPath(), ':'));
 	}
 }